requirements.txt
asgiref==3.7.2
Django==4.2.7
django-quill-editor==0.1.40
Pillow==10.1.0
sqlparse==0.4.4
whitenoise==6.6.0
